+++ autoextract_poet/items.py
@@ -202,6 +202,9 @@ class Product(Item):
     descriptionHtml: Optional[str] = None
     additionalProperty: List[AdditionalProperty] = attr.Factory(list)
     aggregateRating: Optional[Rating] = None
+    color: Optional[str] = None
+    size: Optional[str] = None
+    style: Optional[str] = None
 
     @classmethod
     def from_dict(cls, item: Optional[Dict]):
@@ -437,7 +440,7 @@ class RealEstate(Item):
 
 @attr.s(auto_attribs=True, slots=True)
 class Review(Item):
-    
+
     name: Optional[str] = None
     reviewBody: Optional[str] = None
     reviewRating: Optional[Rating] = None